refactor(cli): extract copy folder fn
